+++ src/main.tsx
@@ -1,6 +1,7 @@
 import { StrictMode } from "react"
 import { createRoot } from "react-dom/client"
 import { ThemeProvider } from "@/components/theme-provider"
+import { Toaster } from "@/components/ui/sonner"
 import { App } from "./App.tsx"
 import "driver.js/dist/driver.css"
 import "./index.css"
@@ -9,6 +10,7 @@ createRoot(document.getElementById("root")!).render(
   <StrictMode>
     <ThemeProvider defaultTheme="system" storageKey="vite-ui-theme">
       <App />
+      <Toaster position="bottom-right" closeButton />
     </ThemeProvider>
   </StrictMode>
 )